refactor(wasm): put wasm related stuff into its own crate
