+++ packages/okp4-logic-bindings/Cargo.toml
@@ -7,6 +7,7 @@ version = "4.1.0"
 [dependencies]
 cosmwasm-std.workspace = true
 form_urlencoded = "1.2.1"
+okp4-wasm.workspace = true
 schemars.workspace = true
 serde-json-wasm.workspace = true
 serde.workspace = true
