+++ packages/okp4-logic-bindings/src/error.rs
@@ -1,21 +1,5 @@
 use std::string::FromUtf8Error;
 use thiserror::Error;
-use url::ParseError;
-
-#[derive(Error, Debug, PartialEq, Eq)]
-pub enum CosmwasmUriError {
-    #[error("{0}")]
-    ParseURI(#[from] ParseError),
-
-    #[error("{0}")]
-    ParseQuery(String),
-
-    #[error("{0}")]
-    SerializeQuery(String),
-
-    #[error("Malformed URI: {0}")]
-    Malformed(String),
-}
 
 #[derive(Error, Debug, PartialEq, Eq)]
 pub enum TermParseError {
