+++ packages/okp4-wasm/Cargo.toml
@@ -1,12 +1,14 @@
 [package]
 authors = ["OKP4"]
 edition = "2021"
-name = "okp4-objectarium-client"
+name = "okp4-wasm"
 version = "4.1.0"
 
 [dependencies]
 cosmwasm-std.workspace = true
-okp4-objectarium.workspace = true
-okp4-wasm.workspace = true
+form_urlencoded = "1.2.1"
 schemars.workspace = true
+serde-json-wasm.workspace = true
 serde.workspace = true
+thiserror.workspace = true
+url = "2.5.0"
