+++ packages/okp4-wasm/Makefile.toml
@@ -1,0 +1,1 @@
+[tasks.schema]
