+++ packages/okp4-wasm/README.md
@@ -1,0 +1,3 @@
+# RDF
+
+Package that holds useful components to manage with `RDF` data, typically reading / writing.
