+++ packages/okp4-wasm/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod error;
+pub mod uri;
